pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.3.0",
    "fastapi>=0.120.2",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
    "python-magic>=0.4.27",
    "apscheduler>=3.11.0",
    "slowapi>=0.1.9",
    "pandas>=2.3.3",
    "twilio>=9.8.5",
    "websockets>=15.0.1",
    "redis>=7.0.1",
    "openpyxl>=3.1.5",
    "reportlab>=4.4.4",
    "sendgrid>=6.12.5",
]
